Se ha ordenado el código en main y libreria
Se ha reordenado todo para dejar las cosas mas claras en el main.
Además en la funcion clickar para evitar tener que clickar dos veces se ha recuperado el foco de la pestaña buscaminas
